fix: namespace route section type
