+++ app/graphql/types/tfl/entities/disruption_type.rb
@@ -3,7 +3,7 @@ module Types
     module Entities
       class DisruptionType < Types::BaseObject
         field :additionalInfo, String, null: true
-        field :affectedRoutes, [Types::RouteSectionType, null: true], null: true
+        field :affectedRoutes, [Types::Tfl::Entities::RouteSectionType, null: true], null: true
         field :affectedStops, [Types::StopPointType, null: true], null: true
         field :category, Types::CategoryType, null: true
         field :categoryDescription, String, null: true
